+++ CHANGES.md
@@ -1,7 +1,7 @@
 # Changes
 
 
-## 0.1.14.dev0 (Work In Progress)
+## 0.1.14 (2019-02-24)
 
 - Delta Peaked calculation fixed in case of consumed memory being negative (released)
 - switch to pynvx's pynvml wrapper